Added (int) to damage, speed, and defense slider in CharacterUpdatePage.xaml.cs

+++ Game/Game/Views/Characters/CharacterUpdatePage.xaml.cs
@@ -138,7 +138,7 @@ namespace Game.Views
         /// <param name="e"></param>
         public void Attack_OnSliderValueChanged(object sender, ValueChangedEventArgs e)
         {
-            AttackLabel.Text = String.Format("{0}", e.NewValue);
+            AttackLabel.Text = String.Format("{0}", (int)e.NewValue);
         }
 
         /// <summary>
@@ -148,7 +148,7 @@ namespace Game.Views
         /// <param name="e"></param>
         public void Defense_OnSliderValueChanged(object sender, ValueChangedEventArgs e)
         {
-            DefenseLabel.Text = String.Format("{0}", e.NewValue);
+            DefenseLabel.Text = String.Format("{0}", (int)e.NewValue);
         }
 
         /// <summary>
@@ -158,7 +158,7 @@ namespace Game.Views
         /// <param name="e"></param>
         public void Speed_OnSliderValueChanged(object sender, ValueChangedEventArgs e)
         {
-            SpeedLabel.Text = String.Format("{0}", e.NewValue);
+            SpeedLabel.Text = String.Format("{0}", (int)e.NewValue);
         }
 
         /// <summary>
